requirements.txt
Pillow == 9.1.1
colorama == 0.4.5
tqdm == 4.64.1
